# project-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Provider Object Storage API
  version: 0.1.0
  description: |
    Initial contract for the implemented control-plane and custom storage API.
    This draft covers the organization membership and role administration
    vertical slice together with provider setup and storage clients.
    Organization management routes require a user JWT and the
    x-organization-id header. Platform routes require a global role with the
    platform permission and do not accept an organization header as scope.
    Storage routes require an API credential in the form keyId.secret and
    enforce namespace and folder-grant boundaries.
servers:
  - url: http://localhost:4000
    description: Local API
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Authentication]
      operationId: login
      description: Sets rotated httpOnly refresh and CSRF cookies.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access token and public user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          description: Origin is not allowed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refresh
      description: Requires the refresh cookie and matching X-CSRF-Token header.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Rotated access and refresh session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/CsrfFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      description: Revokes the refresh session and clears auth cookies.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Session logged out.
        '403':
          $ref: '#/components/responses/CsrfFailed'
  /v1/auth/session:
    get:
      tags: [Authentication]
      operationId: session
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current public user.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/AuthUser'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/providers/{providerId}/test:
    post:
      tags: [Providers]
      operationId: testProvider
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ProviderId'
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                bucketName:
                  type: string
                  minLength: 3
                  maxLength: 255
      responses:
        '200':
          description: Provider test succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderTestResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/ProviderFailure'
  /v1/providers/{providerId}/buckets:
    get:
      tags: [Providers]
      operationId: discoverProviderBuckets
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ProviderId'
        - $ref: '#/components/parameters/OrganizationIdHeader'
      responses:
        '200':
          description: Accessible bucket names.
          content:
            application/json:
              schema:
                type: object
                required: [bucketNames]
                properties:
                  bucketNames:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/ProviderFailure'
  /v1/buckets:
    get:
      tags: [Buckets]
      operationId: listBucketConnections
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      responses:
        '200':
          description: Imported bucket connections.
          content:
            application/json:
              schema:
                type: object
                required: [buckets]
                properties:
                  buckets:
                    type: array
                    items:
                      $ref: '#/components/schemas/BucketConnection'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/buckets/import:
    post:
      tags: [Buckets]
      operationId: importBucket
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BucketImportRequest'
      responses:
        '201':
          description: Bucket imported.
          content:
            application/json:
              schema:
                type: object
                required: [bucket]
                properties:
                  bucket:
                    $ref: '#/components/schemas/BucketConnection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/ProviderFailure'
  /v1/buckets/{bucketId}:
    parameters:
      - $ref: '#/components/parameters/BucketId'
      - $ref: '#/components/parameters/OrganizationIdHeader'
    get:
      tags: [Buckets]
      operationId: getBucketConnection
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Bucket connection.
          content:
            application/json:
              schema:
                type: object
                required: [bucket]
                properties:
                  bucket:
                    $ref: '#/components/schemas/BucketConnection'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Buckets]
      operationId: updateBucketStatus
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              additionalProperties: false
              properties:
                status:
                  type: string
                  enum: [ACTIVE, DISABLED]
      responses:
        '200':
          description: Bucket status updated.
          content:
            application/json:
              schema:
                type: object
                required: [bucket]
                properties:
                  bucket:
                    $ref: '#/components/schemas/BucketConnection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Buckets]
      operationId: deleteBucketConnection
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Bucket connection deleted.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived dashboard access token.
    storageCredential:
      type: http
      scheme: bearer
      bearerFormat: keyId.secret
      description: API credential scoped to one organization and namespace.
    metricsBearer:
      type: http
      scheme: bearer
      description: Deployment-provided metrics scrape token.
  parameters:
    OrganizationIdHeader:
      name: x-organization-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
    ProviderId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BucketId:
      name: bucketId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Must match the non-httpOnly mosp_csrf cookie.
      schema:
        type: string
        minLength: 32
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: Missing, invalid, expired, revoked, or out-of-scope credential.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated principal lacks the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    CsrfFailed:
      description: CSRF cookie/header validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource was not found in the caller's tenant scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource state or idempotency conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate or concurrency limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ProviderFailure:
      description: Provider call failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
...
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 1
          maxLength: 256
          writeOnly: true
    AuthUser:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
    AuthTokenResponse:
      type: object
      required: [accessToken, accessTokenExpiresAt, user]
      properties:
        accessToken:
          type: string
          readOnly: true
        accessTokenExpiresAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/AuthUser'
...
    ProviderTestResult:
      type: object
      required: [ok, bucketNames, capabilities]
      properties:
        ok: { type: boolean }
        bucketNames:
          type: array
          items: { type: string }
        capabilities:
          type: array
          items:
            type: object
            required: [capability, supported]
            properties:
              capability: { type: string }
              supported: { type: boolean }
              details:
                type: object
                additionalProperties: { type: string }
    BucketImportRequest:
      type: object
      required: [bucketName, providerConnectionId]
      additionalProperties: false
      properties:
        bucketName: { type: string, minLength: 3, maxLength: 255 }
        providerConnectionId: { type: string, format: uuid }
    BucketConnection:
      type: object
      required: [id, bucketName, providerConnectionId, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        bucketName: { type: string }
        providerConnectionId: { type: string, format: uuid }
        status: { type: string }
        createdAt: { type: string, format: date-time }
...